refactor(#84): PromptHandler in ChatOrchestrator, CliChatRunner und ServerChatRunner aufteilen
- ChatOrchestrator: gemeinsame, ausgabe-agnostische Logik (ParseToolCalls,
  ExecuteToolCallRoundAsync, ChatOnceAsync, ApplyModelOverride)
- CliChatRunner: CLI-Pfad mit Console-Streaming (aus RunAsync extrahiert)
- ServerChatRunner: Server-Pfad, implementiert IPromptHandler (aus RunServerChatAsync extrahiert)
- PromptHandler.cs gelöscht, Program.cs auf neue Klassen umgestellt
- PromptHandlerTests → ServerChatRunnerTests, alle 171 Tests weiterhin grün
